put history builder in separate file
